+++ src/converter/ButtonConverter.js
@@ -26,7 +26,7 @@ class ButtonConverter extends AbstractConverter {
             image.autoCropTransparent();
 
             const metadata = {
-                nineslice_size: 4,
+                nineslice_size: 5,
                 base_size: [width, height]
             };
 
@@ -67,12 +67,16 @@ class ButtonConverter extends AbstractConverter {
                     [0, 46, 200, 20, [
                         "textures/ui/button_borderless_darkpressednohover",
                         "textures/ui/button_borderless_lightpressednohover",
-                        "textures/ui/disabledButtonNoBorder"
+                        "textures/ui/disabledButtonNoBorder",
+                        "textures/ui/recipe_book_button_borderless_lightpressednohover"
                     ]],
                     [0, 66, 200, 20, [
                         "textures/ui/button_borderless_dark",
                         "textures/ui/button_borderless_light",
-                        "textures/ui/pocket_button_default"
+                        "textures/ui/pocket_button_default",
+                        "textures/ui/recipe_book_button_borderless_light",
+                        "textures/ui/recipe_book_dark_button",
+                        "textures/ui/recipe_book_light_button"
                     ]],
                     [0, 86, 200, 20, [
                         "textures/ui/button_borderless_darkhover",
@@ -80,12 +84,17 @@ class ButtonConverter extends AbstractConverter {
                         "textures/ui/button_borderless_lighthover",
                         "textures/ui/button_borderless_lightpressed",
                         "textures/ui/pocket_button_hover",
-                        "textures/ui/pocket_button_pressed"
+                        "textures/ui/pocket_button_pressed",
+                        "textures/ui/recipe_book_button_borderless_lighthover",
+                        "textures/ui/recipe_book_button_borderless_lightpressed",
+                        "textures/ui/recipe_book_dark_button_pressed",
+                        "textures/ui/recipe_book_light_button_pressed"
                     ]]
                 ],
                 [
                     "textures/ui/focus_border_selected",
-                    "textures/ui/focus_border_white"
+                    "textures/ui/focus_border_white",
+                    "textures/ui/world_screenshot_focus_border"
                 ]
             ],
             [
@@ -94,13 +103,16 @@ class ButtonConverter extends AbstractConverter {
                 [
                     [0, 0, 28, 32, [
                         "textures/ui/pocket_tab_left_side",
+                        "textures/ui/recipe_book_side_toggle_dark",
                         "textures/ui/TabLeftBack",
                         "textures/ui/TabLeftBackBottomMost",
                         "textures/ui/TabLeftBackTopMost",
                         "textures/ui/TabTopBackLeftMost"
                     ]],
                     [28, 0, 28, 32, [
-                        "textures/ui/TabTopBack"
+                        "textures/ui/recipe_back_panel",
+                        "textures/ui/TabTopBack",
+                        "textures/ui/toolbar_background"
                     ]],
                     [140, 0, 28, 32, [
                         "textures/ui/pocket_tab_right_side",
@@ -112,6 +124,7 @@ class ButtonConverter extends AbstractConverter {
                         "textures/ui/XTab"
                     ]],
                     [0, 32, 28, 32, [
+                        "textures/ui/recipe_book_side_toggle_dark_hover",
                         "textures/ui/TabLeftFront",
                         "textures/ui/TabLeftFrontBottomMost",
                         "textures/ui/TabLeftFrontTopMost",
@@ -126,6 +139,31 @@ class ButtonConverter extends AbstractConverter {
                         "textures/ui/TabRightFrontTopMost",
                         "textures/ui/TabTopFrontRightMost",
                         "textures/ui/TabTopFrontRightMostDark"
+                    ]]
+                ],
+                []
+            ],
+            [
+                "textures/gui/container/generic_54.png",
+                256,
+                [
+                    [7, 17, 18, 18, [
+                        "textures/ui/cell_image",
+                        "textures/ui/recipe_book_item_bg"
+                    ]]
+                ],
+                [
+                    "textures/ui/recipe_book_pane_bg",
+                    "textures/ui/recipe_book_touch_cell_selected"
+                ]
+            ],
+            [
+                "textures/gui/container/creative_inventory/tab_item_search.png",
+                256,
+                [
+                    [79, 3, 92, 14, [
+                        "textures/ui/edit_box_indent",
+                        "textures/ui/edit_box_indent_hover"
                     ]]
                 ],
                 []