+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [x]
+- Fix enchanted item glint
+
 ## [2.0.2]
 - Support java v1.15 and bedrock v1.14
 - Convert hotbar, bars, buttons and tabs ui and inventory icons